feat: Cache assets for offline mode
When available in the browser, the site uses a service worker to cache
responses to requests for site assets. Visited pages are rendered when
there is no network access.

initMap function has been refactored, because the Google Maps API related
assets are not cached, so when there is no network connection, we can see
the selected restaurants.